requirements.txt
numpy==1.23.4
nltk==3.7
pandas==1.5.0
tqdm==4.64.1
torch==1.12.1
matplotlib==3.6.1
scikit-learn==1.1.2
gdown==4.7.1
contractions==0.1.73
pyspellchecker==0.7.2